Fix: Properly create RPM-based Storm log dir (usually /opt/storm/logs)
If $log_dir does not point to the default "storm.home"/logs/ directory, we create a symlink
from "storm.home"/logs/ to the actual $log_dir.  This is required because as of Sep 2013
(and Storm 0.9.0-wip21) the log directory is still hardcoded in same places in the Storm
code.  Otherwise we could just supply a custom "storm.home"/logback/cluster.xml config.
See https://groups.google.com/forum/#!topic/storm-user/IKRtIkqQfqc for details.